Add files via upload
The python file is a working "expert system" approach to digit identification, with a success rate of about 75% at the moment. With a bit more optimisation I hope to get past 80%, but I'm not sure I'll do much better.

The ppt file is a first go at some slides for the presentation, though they are too long at the moment. Comments welcome.