Add files via upload
Davids_features_NN.py is code to extract feature data from images and save them to the data files, and/or to run them through a neural net, using the feature data as input.

The npy files are numpy data files, covering the first 50,000 training images and all 10,000 test images, with the features extracted. Each entry is 38 elements.