Add files via upload
Output files from different programs. Note the "vector" and "true_sentiment" files are needed by the basic NN program (which should look for them in the right place in this directory structure). NB The vector file for the main tweet set ("tweets.csv") is too big to upload when uncompressed: you can find it (together with all the others of this type, and their associated sentiment files) in the zip archive.

The airline_, time_zone_ and times_ data files have been put in a single excel workbook, and histograms generated there - see in outputs folder.